% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIVAR_b.r
\name{irf_CIVAR_CB}
\alias{irf_CIVAR_CB}
\title{Impulse response of an estimated CIVAR.}
\usage{
irf_CIVAR_CB(
  res,
  nstep,
  comb,
  irf = c("gen", "chol", "chol1", "gen1", "comb1"),
  runs = 200,
  conf = c(0.05, 0.95)
)
}
\arguments{
\item{res}{: a CIVAR object such as an output of CIVARest.}

\item{nstep}{: the length of the impulse response functions.}

\item{comb}{: a weighting matrix specifying the weights used in the impulse response functions of a global VAR. Its default value is NA for CIVAR(p).}

\item{irf}{: types of the generated impulse response functions.}

\item{runs}{: number of runs used in the the calculation of the bootstrap confidence interval.}

\item{conf}{: a two component vector containing the tail probabilities of the bootstrap confidence interval.}
}
\value{
\preformatted{  an array of dimension (n, n, nstep, 3).
}
}
\description{
This function generates the impulse response functions of an estimated CIVAR model
}
\examples{

res_d = CIVARData(n=4,p=2,T=84,Co=matrix(c(1,1,1,1),4,1)*0,type="none",crk=1)
res_e = CIVARest(res=res_d)
res_e$Summary

IRF_CB = irf_CIVAR_CB(res=res_e, nstep=20, comb=NA, irf = "gen1", runs = 20, conf = c(0.05, 0.95))
IRF_g = IRF_graph(IRF_CB)

}
