% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_b.R
\name{irf_VAR_CB}
\alias{irf_VAR_CB}
\title{Impulse response function of VAR(p)}
\usage{
irf_VAR_CB(
  res,
  nstep,
  comb,
  irf = c("gen", "chol", "chol1", "gen1", "comb1"),
  runs = 200,
  conf = c(0.05, 0.95)
)
}
\arguments{
\item{res}{an object of the output of VARest}

\item{nstep}{length of the impulse response functions}

\item{comb}{a vector containing the weights of a combined policy actions}

\item{irf}{types of impulse response functions}

\item{runs}{number of simulation runs in generating the confidence interval of the impulse response functions}

\item{conf}{confidence levels}
}
\value{
an (n, n, nstep, 3) array containing the impulse response functions
}
\description{
This function generates impulse response functions of an estimated VAR(p)
}
