% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARs_functions_b.R
\name{MRCIVARTestm}
\alias{MRCIVARTestm}
\title{Test restrictions within the cointegration space}
\usage{
MRCIVARTestm(
  res = res,
  H = H,
  h = h,
  phi = phi,
  G = G,
  psi = psi,
  method = c("parametric", "bootstrap")
)
}
\arguments{
\item{res}{An object of the output of CIVARest}

\item{H}{A matrix specifying the restrictions on beta}

\item{h}{A vector specifying the restrictions on beta}

\item{phi}{freely varying parameters in beta}

\item{G}{A matrix specifying the restrictions on alpha}

\item{psi}{freely varying parameters in alpha}

\item{method}{Method used to run the test}
}
\value{
A list containing constrained estimated CIVAR model
}
\description{
This function runs a likelihood ratio test of linear restrictions on \eqn{\alpha} and \eqn{\beta} in a CIVAR model
}
\section{Details}{

This function runs a likelihood ratio test of linear restrictions on alpha and beta in a CIVAR model in the following form:
\deqn{vec(\alpha') = G \psi,  vec(\beta) = H\phi + h}
}

\examples{
Sigma = 1:(4*4*2)
dim(Sigma) = c(4,4,2)
Sigma[,,1] = diag(4)
Sigma[,,2] = diag(4)
p=matrix(0,2,2)
p[,1] = c(2,2)

res_d = MRCIVARDatam(n=4,p=p,T=261,S=2,SESVI=1,TH=0,Sigmao=Sigma,type="const",r=1)
colnames(res_d$Y) = c("w","p","y","r")
res_e = MRCIVARestm1(res=res_d)
res_e$Summary

### case 1 Compare MRCIVARestm1 with AB_MEVARTestm
n = 4; crk = 3
G = diag(12); G2 = list(G,G);  psi = matrix(1,12,1); psi2=list(psi,psi); # No restrictions on alpha
H = diag(12); H2 = H[,-c(1,5,9)]; h = matrix(0,12,1); h[c(1,5,9),1] = c(1,1,1);  phi = matrix(1,9,1)

G2[[1]]\%*\%psi2[[1]]; G2[[2]]\%*\%psi2[[2]]; H2\%*\%phi+h

res_t  = MRCIVARTestm(res=res_d,H=H2,h=h,phi=phi,G=G2,psi=psi2,method=c("parametric"))
res_eR = ABC_MRCIVARestm(res=res_d,H=H2,h=h,phi=phi,G=G2,psi=psi2)

res_eR$LR
res_eR$p_value                                 #### p-value = 1 because thre is no effective restrictions.
res_eR$code
summaryCIVAR(res_eR$VECMR,sname="Z2")


### case 2 test of weak exogeneity of 1st variable
n = 4; crk = 3

G = diag(12); G2 = list(G[,4:12],G[,4:12]);  psi = matrix(1,9,1); psi2=list(psi,psi);


H = diag(12); H2 = H[,-c(1,5,9)]; h = matrix(0,12,1); h[c(1,5,9),1] = c(1,1,1);  phi = matrix(1,9,1)


 G2[[1]]\%*\%psi2[[1]]; G2[[2]]\%*\%psi2[[2]]; H2\%*\%phi+h


res_t = MRCIVARTestm(res=res_d,H=H2,h=h,phi=phi,G=G2,psi=psi2,method=c("parametric"))
res_eR = ABC_MRCIVARestm(res=res_d,H=H2,h=h,phi=phi,G=G2,psi=psi2)

res_eR$LR
res_eR$p_value                                 #### p-value = 1 because thre is no effective restrictions.
res_eR$code
summaryCIVAR(res_eR$VECMR,sname="Z2")
}
