% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GVAR_b.R
\name{W2Wmat}
\alias{W2Wmat}
\title{Transformation of a weighting vector to a weighting matrix}
\usage{
W2Wmat(W)
}
\arguments{
\item{W}{the weighting vector}
}
\value{
the corresponding weighting matrix
}
\description{
Transformation of a weighting vector to a weighting matrix
}
